test(egl): add unit tests for EGL security fixes
- EglGetProcAddressImpl003: test nullptr procname (#27)
- EglGetConfigsImpl004: test null numConfig early return (#25)
- EglGetCurrentDisplayImpl001/002: no context + post-destroy TLS clear (#28)
- EglValidateEglContext001/002: null context + valid then destroyed (#28)
- EvictionSmallShaderSizeTest: eviction with blobSizeMax_=1 (#16)
- EvictionDefaultSizeTest: eviction with 650 blobs at default limit
- SetCacheShaderSizeBoundaryTest: boundary values 0/MAX/MAX+1 (#10)



🤖 AI[100%] 🔧 Human Fixed[0%] 🧑 Human[0%] 👌 AI Adopted[100%]

+++ frameworks/opengl_wrapper/src/EGL/egl_system_layers_manager.cpp
@@ -153,7 +153,7 @@ std::vector<std::string> EglSystemLayersManager::GetStringVectorFromJson(const J
         if (i.isString()) {
             stringVector.push_back(i.asString());
         } else {
-            WLOGD("json item is not a string, type %{public}d", i.type());
+            WLOGD("json item is not a string, type %{public}d", static_cast<int>(i.type()));
         }
     }
